WIP - BROKEN!!! -Trying to get the Fav Icon to work in vertical toolbar
The Fav Icon uses tricks because of limitations in the xul files (not
drawing as specifications say they should).  This checkin works in the
vert toolbar, but only works in the regular toolbar after it has been
removed and re-added.  Maybe try again when Mozilla gets things fixed?
Didn't want to lose my work.